requirements.txt
# Core
fastapi>=0.115
uvicorn>=0.34
neo4j>=5.26
pydantic>=2.0
httpx>=0.28

# Embeddings & Extraction
google-genai>=1.0
graphiti-core[anthropic,google-genai]>=0.28

# Graph RAG (Phase 3)
neo4j-graphrag>=1.0

# Content extraction
readability-lxml>=0.8
beautifulsoup4>=4.12

# Staging & cost tracking
psycopg2-binary>=2.9
asyncpg>=0.29

# Discord watcher (ingestion daemon)
discord.py>=2.3
anthropic>=0.40
youtube-transcript-api>=0.6
lxml>=5.0

# MCP server
mcp>=1.0

# Testing
pytest>=7.0
pytest-asyncio>=0.24
